move APIFnMod to action_update
